// packages/frontend/src/generated/types/CreateOwnershipTransferRequest.ts
/**
 * Generated by Kubb (https://kubb.dev/).
 * Do not edit manually.
 */

import type { OwnershipTransferItemCompositeRequest } from './OwnershipTransferItemCompositeRequest.ts'

export interface CreateOwnershipTransferRequest {
  /**
   * @type string, date-time
   */
  date: string
  /**
   * @type array
   */
  items: OwnershipTransferItemCompositeRequest[]
}