// packages/frontend/src/generated/types/OwnershipTransferItemCompositeRequest.ts
/**
 * Generated by Kubb (https://kubb.dev/).
 * Do not edit manually.
 */

export interface OwnershipTransferItemCompositeRequest {
  /**
   * @type string
   */
  amount: string
  /**
   * @type string, uuid
   */
  fromContractorId: string
  /**
   * @type string, uuid
   */
  productId: string
  /**
   * @type string, uuid
   */
  storageId: string
  /**
   * @type string, uuid
   */
  toContractorId: string
}