// packages/frontend/src/generated/zod/createOwnershipTransferRequestSchema.ts
/**
 * Generated by Kubb (https://kubb.dev/).
 * Do not edit manually.
 */

import type { ToZod } from '../.kubb/ToZod.ts'
import type { CreateOwnershipTransferRequest } from '../types/CreateOwnershipTransferRequest.ts'
import * as z from 'zod'
import { ownershipTransferItemCompositeRequestSchema } from './ownershipTransferItemCompositeRequestSchema.ts'

export const createOwnershipTransferRequestSchema = z.object({
  date: z.string().datetime(),
  items: z.array(z.lazy(() => ownershipTransferItemCompositeRequestSchema)),
}) as unknown as ToZod<CreateOwnershipTransferRequest>