// packages/frontend/src/generated/zod/ownershipTransferItemCompositeRequestSchema.ts
/**
 * Generated by Kubb (https://kubb.dev/).
 * Do not edit manually.
 */

import type { ToZod } from '../.kubb/ToZod.ts'
import type { OwnershipTransferItemCompositeRequest } from '../types/OwnershipTransferItemCompositeRequest.ts'
import * as z from 'zod'

export const ownershipTransferItemCompositeRequestSchema = z.object({
  amount: z.string(),
  fromContractorId: z.string().uuid(),
  productId: z.string().uuid(),
  storageId: z.string().uuid(),
  toContractorId: z.string().uuid(),
}) as unknown as ToZod<OwnershipTransferItemCompositeRequest>